create slice.js using reduxjs/toolkit to process services and reducer at the same codes
